Phase 1: data-model spine for skill tiers + inventory consolidation
Lands the type-level changes from plans/skill_inventory_landing.md
Phase 1. Behavior changes are limited to what types force; Phase 2
rewrites mechanics under the new shape.

- SkillTier enum (Untrained/Trained/Expert) replaces int skill levels
- ItemInstance.IsEquipped flag replaces dedicated EquippedGear slots
- PlayerState.Pack absorbs Haversack; HaversackCapacity removed
- StartingPackSlots 3 -> 8
- ItemType.Token retired (8 token defs deleted); bandage def removed
- medical_kit added as Tool (auto-cure behavior wired in Phase 2)
- EquippedGear.cs deleted; EquipmentShim.cs added as transitional
  layer for Equipment.X = item writes (Phase 2 removes callers)
- SkillChecks token-bonus methods stubbed to zero (Phase 2 deletes)
- Tests translated to tiers and pack-only inventory

465 passed / 3 skipped / 1 pre-existing Combat.Tests failure
unrelated to this change. Encounter.sln builds clean.

+++ lib/Game/Choices.cs
@@ -87,7 +87,7 @@ public static class Choices
                     var skill = Skills.FromScriptName(tokens[1]);
                     if (skill != null && int.TryParse(tokens[2], out var target))
                     {
-                        var skillLevel = state.Skills.GetValueOrDefault(skill.Value);
+                        var skillLevel = (int)state.Skills.GetValueOrDefault(skill.Value);
                         var itemBonus = SkillChecks.GetItemBonus(skill.Value, state, balance);
                         var total = skillLevel + itemBonus;
                         passed = total >= target;
